Extracted Messaging caching in it's own class
